This is the Aurora.jl documentation

Welcome to the documentation page

Work in progress

This project is still in progress. Be aware that the documentation might not be up to date.

Aurora.setup — Function

This is the setup function, loading the atmosphere, the energy grid, the collision cross-sections, ...
It calls a lot of functions from the original MATLAB code.

Inputs

top_altitude: the altitude, in km, for the top of the ionosphere in our simulation

θ_lims: range of angles for the limits of our electron beams, i.e 180:-10:0

E_max: upper limit for the energy grid (in eV)